Remove account name field from AddAccountView.
It is an extra information for the user to have to enter and also
redundant because we can use email address to identify an account. I did
not remove it from the database so we can revive it if we decide to do so
later. For now, the email address populates both the account name and the
email address field in the database.