Fix preloaded email not being marked as 'read'
Sometimes an email whose body has been downloaded already may be marked as
unread. I.e., the user explicitly marked the message as unread after
reading it. This commit fixes the situation where the \Seen flag was only
getting set when the message body is freshly being downloaded. Now the
flag is set even when you are reading a predownloaded email.